feat(bluetooth): raise autotype script cap from 255 to 998
Widen the keyboard autotype script count and every script index from
uint8_t to uint16_t across the bluetooth, lcd, and ai components so more
than 255 scripts can be stored, listed, and typed. Category counts and
indices stay uint8_t (capped at 50).

998 is the hard ceiling: script selections encode as
BLUETOOTH_SCRIPT_OFFSET(1000) + BT_NUM_KEYBOARD_BASE(2) + idx into the
uint16_t command queue and must stay below the special-command block at
BLUETOOTH_SCRIPT_PRESENTATION_START(2000). The ~236 KiB NVS partition is
the practical limit (a few hundred small scripts); saves and imports
return an error rather than corrupting when flash fills.

- Store the script count as NVS u16; migrate the legacy u8 value in
  bluetooth_portal_script_count_get_nvs() via nvs_set_u16 (no erase, so
  the rewrite is crash-safe), so upgraded devices keep their scripts
  without re-import.
- script_one_post(): write label/body/cat before bumping the count last,
  so a failed write (e.g. NVS full) never leaves a counted-but-empty slot.
- Scale the AI custom/credential lookup catalog buffers from a fixed 18KB
  to size-from-cap macros, so all saved scripts are offered to the model.
- LCD script_indices[] widened to uint16_t; "invalid row" sentinel moved
  from 255 to UINT16_MAX. Portal index input max and JS guard updated.

+++ components/bluetooth/src/bluetooth_task.c
@@ -337,7 +337,7 @@ static void bluetooth_task(void *arg)
                 // Menu has 2 fixed rows before user scripts:
                 // So the first user script is menu_idx == 2 -> NVS index 0
                 if (menu_idx >= BT_NUM_KEYBOARD_BASE) {
-                    uint8_t script_idx = (uint8_t)(menu_idx - BT_NUM_KEYBOARD_BASE); // 0-based NVS slot
+                    uint16_t script_idx = (uint16_t)(menu_idx - BT_NUM_KEYBOARD_BASE); // 0-based NVS slot
 
                     size_t blen = 0;
